move out TransformerEncoder to separate file
