+++ metaseq/model_parallel/models/transformer.py
@@ -9,7 +9,8 @@ from metaseq.model_parallel.modules import (
     ModelParallelTransformerDecoderLayer,
     ModelParallelTransformerEncoderLayer,
 )
-from metaseq.models.transformer import TransformerDecoder, TransformerEncoder
+from metaseq.models.transformer import TransformerDecoder
+from metaseq.models.transformer_encoder import TransformerEncoder
 
 try:
     from megatron.mpu import (
